more work on parallel passes
