+++ packages/native-core/src/tree.rs
@@ -1,20 +1,11 @@
 use core::panic;
-use std::hash::BuildHasherDefault;
-use std::sync::atomic::{AtomicU64, Ordering};
-
-use crossbeam_deque::{Injector, Stealer, Worker};
-use dashmap::DashSet;
-use dioxus_core::ScopeId;
-use dioxus_html::u;
 use parking_lot::lock_api::RawMutex as _;
-use parking_lot::{Mutex, RawMutex, RwLock, RwLockWriteGuard};
-use rustc_hash::{FxHashMap, FxHashSet, FxHasher};
+use parking_lot::{RawMutex, RwLock};
 use slab::Slab;
 use std::cell::UnsafeCell;
 use std::collections::VecDeque;
 use std::marker::PhantomData;
 use std::sync::Arc;
-use std::thread;
 
 #[derive(Hash, PartialEq, Eq, Clone, Copy, Debug, PartialOrd, Ord)]
 pub struct NodeId(pub usize);
@@ -109,7 +100,7 @@ pub trait TreeView<T>: Sized {
             // Safety: No node has itself as a parent.
             (*mut_ptr)
                 .get_mut(id)
-                .map(|node| (node, (*mut_ptr).parent_mut(id).map(|parent| parent)))
+                .map(|node| (node, (*mut_ptr).parent_mut(id)))
         }
     }
 
@@ -807,250 +798,4 @@ fn traverse_depth_first() {
         assert_eq!(*node, node_count);
         node_count += 1;
     });
-}
-
-#[test]
-fn traverse_breadth_first() {
-    let mut tree = Tree::new(0);
-    let parent = tree.root();
-    let child1 = tree.create_node(1);
-    tree.add_child(parent, child1);
-    let grandchild1 = tree.create_node(3);
-    tree.add_child(child1, grandchild1);
-    let child2 = tree.create_node(2);
-    tree.add_child(parent, child2);
-    let grandchild2 = tree.create_node(4);
-    tree.add_child(child2, grandchild2);
-
-    let mut node_count = 0;
-    tree.traverse_breadth_first(move |node| {
-        assert_eq!(*node, node_count);
-        node_count += 1;
-    });
-}
-
-#[derive(PartialEq, Eq, Hash, Clone, Copy)]
-pub struct PassId(u64);
-
-pub trait UpwardPass<T> {
-    fn pass_id(&self) -> PassId;
-    fn dependancies(&self) -> &'static [PassId];
-    fn dependants(&self) -> &'static [PassId];
-    fn upward_pass(&self, node: &mut T, parent: Option<&mut T>) -> bool;
-}
-
-pub trait DownwardPass<T> {
-    fn pass_id(&self) -> PassId;
-    fn dependancies(&self) -> &'static [PassId];
-    fn dependants(&self) -> &'static [PassId];
-    fn downward_pass<'a>(
-        &self,
-        node: &mut T,
-        children: &mut dyn Iterator<Item = &'a mut T>,
-    ) -> bool;
-}
-
-pub trait NodePass<T> {
-    fn pass_id(&self) -> PassId;
-    fn dependancies(&self) -> &'static [PassId];
-    fn dependants(&self) -> &'static [PassId];
-    fn node_pass(&self, node: &mut T) -> bool;
-}
-
-pub enum AnyPass<T> {
-    Upward(Box<dyn UpwardPass<T> + Send + Sync>),
-    Downward(Box<dyn DownwardPass<T> + Send + Sync>),
-    Node(Box<dyn NodePass<T> + Send + Sync>),
-}
-
-impl<T> AnyPass<T> {
-    fn pass_id(&self) -> PassId {
-        match self {
-            Self::Upward(pass) => pass.pass_id(),
-            Self::Downward(pass) => pass.pass_id(),
-            Self::Node(pass) => pass.pass_id(),
-        }
-    }
-
-    fn dependancies(&self) -> &'static [PassId] {
-        match self {
-            Self::Upward(pass) => pass.dependancies(),
-            Self::Downward(pass) => pass.dependancies(),
-            Self::Node(pass) => pass.dependancies(),
-        }
-    }
-
-    fn dependants(&self) -> &'static [PassId] {
-        match self {
-            Self::Upward(pass) => pass.dependants(),
-            Self::Downward(pass) => pass.dependants(),
-            Self::Node(pass) => pass.dependants(),
-        }
-    }
-}
-
-type FxDashSet<T> = dashmap::DashSet<T, BuildHasherDefault<FxHasher>>;
-type FxDashMap<K, V> = dashmap::DashMap<K, V, BuildHasherDefault<FxHasher>>;
-
-#[derive(Default)]
-struct DirtyNodeStates {
-    dirty: FxDashMap<NodeId, Vec<AtomicU64>>,
-}
-
-impl DirtyNodeStates {
-    fn new(starting_nodes: FxHashMap<NodeId, FxDashSet<PassId>>) -> Self {
-        let mut this = Self::default();
-        for (node, nodes) in starting_nodes {
-            for pass_id in nodes {
-                this.insert(pass_id, node);
-            }
-        }
-        this
-    }
-
-    fn insert(&self, pass_id: PassId, node_id: NodeId) {
-        let pass_id = pass_id.0;
-        let index = pass_id / 64;
-        let bit = pass_id % 64;
-        let encoded = 1 << bit;
-        if let Some(dirty) = self.dirty.get(&node_id) {
-            if let Some(atomic) = dirty.get(index as usize) {
-                atomic.fetch_or(encoded, Ordering::Relaxed);
-            } else {
-                drop(dirty);
-                let mut write = self.dirty.get_mut(&node_id).unwrap();
-                write.resize_with(index as usize + 1, || AtomicU64::new(0));
-                write[index as usize].fetch_or(encoded, Ordering::Relaxed);
-            }
-        } else {
-            self.dirty.insert(node_id, vec![AtomicU64::new(encoded)]);
-        }
-    }
-
-    fn all_dirty(&self, pass_id: PassId) -> impl Iterator<Item = NodeId> + '_ {
-        let pass_id = pass_id.0;
-        let index = pass_id / 64;
-        let bit = pass_id % 64;
-        let encoded = 1 << bit;
-        self.dirty.iter().filter_map(move |entry| {
-            let node_id = entry.key();
-            let dirty = entry.value();
-            if let Some(atomic) = dirty.get(index as usize) {
-                if atomic.load(Ordering::Relaxed) & encoded != 0 {
-                    Some(*node_id)
-                } else {
-                    None
-                }
-            } else {
-                None
-            }
-        })
-    }
-}
-
-pub fn resolve_passes<T>(
-    tree: &mut impl TreeView<T>,
-    starting_nodes: FxHashMap<NodeId, FxDashSet<PassId>>,
-    mut passes: Vec<AnyPass<T>>,
-) {
-    let dirty_nodes: Arc<DirtyNodeStates> = Arc::new(DirtyNodeStates::new(starting_nodes));
-    let global = Injector::default();
-
-    let core_count = thread::available_parallelism()
-        .map(|c| c.get())
-        .unwrap_or(1);
-    let workers: Vec<Worker<NodeId>> = (0..core_count).map(|_| Worker::new_fifo()).collect();
-    let stealers: Vec<_> = workers.iter().map(|w| w.stealer()).collect();
-    let shared_view = SharedView::new(tree);
-    let mut resolved_passes: FxHashSet<PassId> = FxHashSet::default();
-    let current_pass: Arc<RwLock<Option<AnyPass<T>>>> = Arc::new(RwLock::new(None));
-    thread::scope(|s| {
-        let global = &global;
-        let stealers = &stealers;
-        for (_, w) in (0..core_count).zip(workers.into_iter()) {
-            let mut shared_view = shared_view.clone();
-            let current_pass = current_pass.clone();
-            let dirty_nodes = dirty_nodes.clone();
-            s.spawn(move || {
-                while let Some(current_pass) = &*current_pass.read() {
-                    match current_pass {
-                        AnyPass::Upward(pass) => {
-                            while let Some(id) = find_task(&w, global, stealers) {
-                                let (node, parent) = shared_view.node_parent_mut(id).unwrap();
-                                if pass.upward_pass(node, parent) {
-                                    if let Some(id) = shared_view.parent_id(id) {
-                                        for dependant in pass.dependants() {
-                                            dirty_nodes.insert(*dependant, id);
-                                        }
-                                        w.push(id);
-                                    }
-                                }
-                            }
-                        }
-                        AnyPass::Downward(pass) => {
-                            while let Some(id) = find_task(&w, global, stealers) {
-                                let (node, mut children) =
-                                    shared_view.parent_child_mut(id).unwrap();
-                                if pass.downward_pass(node, &mut children) {
-                                    drop(children);
-                                    for id in shared_view.children_ids(id).unwrap() {
-                                        for dependant in pass.dependants() {
-                                            dirty_nodes.insert(*dependant, *id);
-                                        }
-                                        w.push(*id);
-                                    }
-                                }
-                            }
-                        }
-                        AnyPass::Node(pass) => {
-                            while let Some(id) = find_task(&w, global, stealers) {
-                                let node = shared_view.get_mut(id).unwrap();
-                                if pass.node_pass(node) {
-                                    for dependant in pass.dependants() {
-                                        dirty_nodes.insert(*dependant, id);
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            });
-        }
-        while !passes.is_empty() {
-            for i in 0..passes.len() {
-                if passes[i]
-                    .dependancies()
-                    .iter()
-                    .all(|id| resolved_passes.contains(id))
-                {
-                    let pass = passes.remove(i);
-                    let pass_id = pass.pass_id();
-                    for node in dirty_nodes.all_dirty(pass_id) {
-                        global.push(node);
-                    }
-                    resolved_passes.insert(pass_id);
-                    break;
-                }
-            }
-        }
-        *current_pass.write() = None;
-    });
-}
-
-fn find_task<T>(local: &Worker<T>, global: &Injector<T>, stealers: &[Stealer<T>]) -> Option<T> {
-    // Pop a task from the local queue, if not empty.
-    local.pop().or_else(|| {
-        // Otherwise, we need to look for a task elsewhere.
-        std::iter::repeat_with(|| {
-            // Try stealing a batch of tasks from the global queue.
-            global
-                .steal_batch_and_pop(local)
-                // Or try stealing a task from one of the other threads.
-                .or_else(|| stealers.iter().map(|s| s.steal()).collect())
-        })
-        // Loop while no task was stolen and any steal operation needs to be retried.
-        .find(|s| !s.is_retry())
-        // Extract the stolen task, if there is one.
-        .and_then(|s| s.success())
-    })
 }
